[feat] Moved app code to SourcesService.ts and started express in index
